// src/generated/models/Dga_app_notificationsesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dga_app_notificationsesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dga_app_notificationsesstatecode = keyof typeof Dga_app_notificationsesstatecode;
export const Dga_app_notificationsesstatuscode = {
  1: 'Open',
  2: 'Inactive',
  576610001: 'Closed'
} as const;
export type Dga_app_notificationsesstatuscode = keyof typeof Dga_app_notificationsesstatuscode;

export interface Dga_app_notificationsesBase {
  dga_app_notificationsid: string;
  dga_notification_id?: string;
  "dga_notification_recipient_team@odata.bind"?: string;
  dga_notification_text?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  "OwningBusinessUnit@odata.bind"?: string;
  statecode: Dga_app_notificationsesstatecode;
  statuscode?: Dga_app_notificationsesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dga_app_notificationses extends Dga_app_notificationsesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dga_notification_recipient_teamname?: string;
  dga_notification_recipient_teamyominame?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  dga_notification_recipient_team?: object;
  _dga_notification_recipient_team_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}